# src/main/resources/static/api.yaml
openapi: "3.0.2"
info:
  description: "This is the [Scoold](https://scoold.com) RESTful API. You need to generate an [API key](/admin) from the 'Administration' page to access the API."
  version: "1.0.0"
  title: "Scoold API"
...
paths:
...
  /api/users/{id}/questions:
    get:
      summary: Returns a list of questions created by a user
      tags:
        - users
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to retrieve
          schema:
            type: string
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/sortby'
        - $ref: '#/components/parameters/desc'
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
        '404':
          description: Not found
...
  /api/tags/{id}/questions:
    get:
      summary: List questions tagged with a specific tag
      tags:
        - tags
      security:
      - scoold_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the tag
          schema:
            type: string
      responses:
        '200':
          description: Tag object returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
...
components:
  schemas:
    Post:
      required:
        - id
        - title
      properties:
        id:
          type: integer
          format: int64
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of when the post was created
        type:
          type: string
          enum:
            - "question"
            - "reply"
            - "sticky"
          description: Can be one of `question`, `reply`, `sticky`
        name:
          type: string
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        creatorid:
          type: string
          description: The `id` of the author of this post (an existing user)
        location:
          type: string
          description: The location name associated with this post
        latlng:
          type: string
          description: Coordinates of the location in the format `"43.26,25.36"`
        address:
          type: string
          description: The location address associated with this post
        lastactivity:
          type: integer
          format: int64
        lastedited:
          type: integer
          format: int64
        lasteditby:
          type: string
        answerid:
          type: string
        revisionid:
          type: string
        closerid:
          type: string
        wiki:
          type: boolean
        space:
          type: string
          description: Selects a space or returns questions in the default space
        comments:
          $ref: '#/components/schemas/Comments'
        children:
          $ref: '#/components/schemas/Posts'
        similar:
          $ref: '#/components/schemas/Posts'
    Posts:
      type: array
      items:
        $ref: '#/components/schemas/Post'
...
    Comment:
      type: object
      properties:
        id:
          type: string
        creatorid:
          type: string
        parentid:
          type: string
        timestamp:
          type: integer
          format: int64
        comment:
          type: string
        hidden:
          type: boolean
        authorName:
          type: string
    Comments:
      type: array
      items:
        $ref: '#/components/schemas/Comment'
...
  parameters:
    limit:
      name: limit
      in: query
      description: How many items to return at one time (default 30)
      required: false
      schema:
        type: integer
        format: int32
        default: 30
    page:
      name: page
      in: query
      description: Page number (paginate through results)
      required: false
      schema:
        type: integer
        format: int32
        default: 1
    sortby:
      name: sortby
      in: query
      description: Field name on which results will be sorted
      required: false
      schema:
        type: string
        default: "timestamp"
    desc:
      name: desc
      in: query
      description: Descending (true) or ascending (false) sort order
      required: false
      schema:
        type: boolean
        default: true
  securitySchemes:
    scoold_auth:
      description: Don't have an API key? Generate one on the [Administration page](/admin).
      type: http
      scheme: bearer
      bearerFormat: JWT